Moved files into dedicated folders
